utils: move task functions to asyncio module
